attemp to fix interaction
